separate component for pagination control
